requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
caldav==0.8.0
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.3
lxml==4.6.3
python-dateutil==2.8.2
python-dotenv==0.19.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.7
vobject==0.9.6.1
 